chore: rework pagination to current standard
