+++ src/js/base/pagination/PaginationContent.tsx
@@ -1,0 +1,8 @@
+import { cn } from "@utils/utils";
+import * as React from "react";
+
+export const PaginationContent = React.forwardRef<HTMLUListElement, React.ComponentProps<"ul">>(
+    ({ className, ...props }, ref) => (
+        <ul ref={ref} className={cn("flex flex-row items-center gap-1", className)} {...props} />
+    )
+);
