+++ src/js/base/pagination/__tests__/Pagination.test.tsx
@@ -1,9 +1,9 @@
+import { Pagination } from "@base";
 import { screen } from "@testing-library/react";
+import { renderWithProviders } from "@tests/setupTests.jsx";
 import React from "react";
 import { MemoryRouter } from "react-router-dom";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { renderWithProviders } from "../../../tests/setupTests";
-import { Pagination } from "../Pagination";
 
 describe("<Pagination />", () => {
     let props;
